Substantial refactoring of babilim.
Moved towards the Callback Schema from Keras to allow for greater flexibility.

This lead to a new updated example and a simplification of the metric + loss. Metric and Loss are merged into a single Loss class (which can compute and log non loss values as well). This encourages a separation between metrics that are computed during training to see how the approach performs and the final evaluation metrics which are computed once on the best model.

Also the data package was refactored to be better structured and more focused on the important parts of the api.